Macro cortex_m_rt::default_handler

macro_rules! default_handler {
    ($path:path) => { ... };
}

This macro lets you override the default exception handler

The first and only argument to this macro is the path to the function that will be used as the default handler. That function must have signature fn()

Examples

default_handler!(foo::bar);

mod foo {
    pub fn bar() {
        ::cortex_m::asm::bkpt();
        loop {}
    }
}