Struct cortex_m_rt::ExceptionFrame

#[repr(C)]
pub struct ExceptionFrame { /* private fields */ }

Registers stacked (pushed onto the stack) during an exception.

Implementations

impl ExceptionFrame

pub fn r0(&self) -> u32

Returns the value of (general purpose) register 0.

pub fn r1(&self) -> u32

Returns the value of (general purpose) register 1.

pub fn r2(&self) -> u32

Returns the value of (general purpose) register 2.

pub fn r3(&self) -> u32

Returns the value of (general purpose) register 3.

pub fn r12(&self) -> u32

Returns the value of (general purpose) register 12.

pub fn lr(&self) -> u32

Returns the value of the Link Register.

pub fn pc(&self) -> u32

Returns the value of the Program Counter.

pub fn xpsr(&self) -> u32

Returns the value of the Program Status Register.

pub unsafe fn set_r0(&mut self, value: u32)

Sets the stacked value of (general purpose) register 0.

Safety

This affects the r0 register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_r1(&mut self, value: u32)

Sets the stacked value of (general purpose) register 1.

Safety

This affects the r1 register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_r2(&mut self, value: u32)

Sets the stacked value of (general purpose) register 2.

Safety

This affects the r2 register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_r3(&mut self, value: u32)

Sets the stacked value of (general purpose) register 3.

Safety

This affects the r3 register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_r12(&mut self, value: u32)

Sets the stacked value of (general purpose) register 12.

Safety

This affects the r12 register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_lr(&mut self, value: u32)

Sets the stacked value of the Link Register.

Safety

This affects the lr register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_pc(&mut self, value: u32)

Sets the stacked value of the Program Counter.

Safety

This affects the pc register of the preempted code, which must not rely on it getting restored to its previous value.

pub unsafe fn set_xpsr(&mut self, value: u32)

Sets the stacked value of the Program Status Register.

Safety

This affects the xPSR registers (IPSR, APSR, and EPSR) of the preempted code, which must not rely on them getting restored to their previous value.